// server/src/api/types.ts
export interface AuthOptional {
  Authorization?: string;
}

export interface AuthRequired {
  Authorization: string;
}

export interface RegisterUserRequest {
  email: string;
  password: string;
  name: string;
  role: Role;
}
// ...
export type Role = "client" | "realtor" | "admin";
// ...
export type RegisterUser_Response = {
  kind: "success";
  data: string;
} | {
  kind: "unauthorized";
  data: string;
} | {
  kind: "failure";
  data: string;
};